Add the core functionality of a file parser
- Read the file
- Perform both simple string match as well as a regular expression match
- If it all works out, then parse the file with SourceKit
